Fixed type error

+++ Classes/TimeZonePickerDelegate.swift
@@ -9,5 +9,5 @@
 import Foundation
 
 public protocol TimeZonePickerDelegate: class {
-    func timeZonePicker(_ timeZonePicker: TimeZonePickerViewController, didSelectTimeZone timeZone: TimeZone, _ lat: Int, _ lng: Int)
+    func timeZonePicker(_ timeZonePicker: TimeZonePickerViewController, didSelectTimeZone timeZone: TimeZone, _ lat: Double, _ lng: Double)
 }
